add iconpicker

+++ app/Http/Controllers/SejarahController.php
@@ -19,23 +19,26 @@ class SejarahController extends Controller
     public function store(Request $request)
     {
         $messages = ['required' => 'field tidak boleh kosong', 'image' => 'format file harus image'];
-        $request->validate(['foto' => 'required|image', 'isi' => 'required', 'tanggal' => 'required'], $messages);
+        $request->validate(['foto' => 'image', 'isi' => 'required', 'tanggal' => 'required'], $messages);
 
-        $file = $request->file('foto');
-        $path = Storage::exists($request->path_foto);
-        if($path){
-            Storage::delete($path);
+        $namaFile = '';
+        if ($request->hasFile('foto')) {
+            $file = $request->file('foto');
+            $path = Storage::exists($request->path_foto);
+            if($path){
+                Storage::delete($path);
+            }
+            $namaFile = Storage::putFile('upload/sejarah', $file);
+            $data['foto'] = $namaFile;
         }
-        $namaFile = Storage::putFile('upload/sejarah', $file);
 
         $data = [
-            'foto' => $namaFile,
-            'isi' => $request->isi,
+            'isi' => htmlspecialchars_decode($request->isi),
             'tanggal' => $request->tanggal,
             'status' => $request->status
         ];
 
-        SejarahPerushaan::updateOrCreate($data);
+        SejarahPerushaan::updateOrCreate(['id' => $request->id], $data);
 
         return redirect()->back()->withSuccess('Sejarah perusahaan berhasil di input');
     }
